Create worksheet9 input and output.hs
Introduction
This worksheet aims to introduce you to Haskell’s input/output mechanism. Begin by downloading
the Week9.hs file from the unit web-site which includes some definitions from the
lecture - experiment with these definitions before moving onto the exercises.
Exercises
1. Write a greeting program that asks the user for his or her name, and then outputs
a personalised greeting. (For example, on input “Sam”, the program should display
“Hello, Sam”.
2. Write an addTwoNumbers program that reads two integers, each on a separate line, and
displays their sum.
3. Write a copyFile program that copies a text-file. The program should ask the user for
the name of the file to copy and the name that the copy should be called.
4. Write a program that allows the user to build up a list of strings by entering them in
one at a time, and displays the list after every step. The program should exit when the
user enters an empty string. An example execution is as follows:
Enter a line: hello
List is now ["hello"]
Enter a line: world
List is now ["hello","world"]
Enter a line:
Hint: use two definitions - a recursive
buildList :: [String] -> IO ()
function that builds up a list step-by-step, and another function listBuilder of type IO
() that starts off the computation by calling buildList with a suitable initial parameter
value.
5. Write a program that reads an integer n from the user, then reads n integers (on separate
lines), and finally displays the sum of the n numbers read. (Again, use two
functions.)
6. [harder] TheWeek8.hs file contains two pure functions for converting between Fahrenheit
and Celsius temperatures. Using these functions, write a program that allows the
user to enter any number of temperatures, and then displays the average of all the
temperatures entered. After entering each value, the user should be prompted to enter
either an “f” or a “c” to say whether the entered value was in Fahrenheit or Celsius.
The user should type a special value (e.g. the empty string or “stop”) to finish. The
average temperature should be outputted in both Fahrenheit and Celsius.